Introduce new module for the trait. Rename `client` module
