+++ crates/e2e/src/backend.rs
@@ -1,0 +1,4 @@
+use jsonrpsee::core::async_trait;
+
+#[async_trait]
+pub trait E2EBackend {}
