+++ crates/e2e/src/lib.rs
@@ -19,50 +19,33 @@
     html_favicon_url = "https://use.ink/crate-docs/favicon.png"
 )]
 
+mod backend;
 mod builders;
-mod client;
 mod contract_results;
 mod default_accounts;
 mod error;
 pub mod events;
 mod node_proc;
+mod subxt_client;
 mod xts;
 
-pub use client::{
-    CallBuilderFinal,
-    Client,
-    Error,
-};
+pub use backend::E2EBackend;
 pub use contract_results::{
-    CallDryRunResult,
-    CallResult,
-    InstantiationResult,
-    UploadResult,
+    CallDryRunResult, CallResult, InstantiationResult, UploadResult,
 };
 pub use default_accounts::*;
 pub use ink_e2e_macro::test;
-pub use node_proc::{
-    TestNodeProcess,
-    TestNodeProcessBuilder,
-};
+pub use node_proc::{TestNodeProcess, TestNodeProcessBuilder};
 pub use sp_core::H256;
 pub use sp_keyring::AccountKeyring;
-pub use subxt::{
-    self,
-    tx::PairSigner,
-};
+pub use subxt::{self, tx::PairSigner};
+pub use subxt_client::{CallBuilderFinal, Client, Error};
 pub use tokio;
 pub use tracing_subscriber;
 
-use pallet_contracts_primitives::{
-    ContractExecResult,
-    ContractInstantiateResult,
-};
+use pallet_contracts_primitives::{ContractExecResult, ContractInstantiateResult};
 use sp_core::sr25519;
-use std::{
-    cell::RefCell,
-    sync::Once,
-};
+use std::{cell::RefCell, sync::Once};
 use xts::ContractsApi;
 
 pub use subxt::PolkadotConfig;
